Chore: moved internal editing extensions and change List deltas to a single delta
